AUDIO: Code bug fix for At command to set the PMU gain [Defect] <MobC00194784>
[Problem]
If AT command at*maudmode=100,3,268,<gain_in_Q13.2> send a gain which is negative value, the code will interpret it as a big positive value.

[Solution]
gain in the code should be defined as short, instead of int.

[Reviewers]
Kernel team
Audio teams.

[Testing]
    [Test Cases]
    Do a music playback to IHF speaker and set the following command to adjust its gain.
    at*maudmode=100, 3, 268, 65535  //-0.25dB.

    [Label/CL# Synced To]
HEAD of sdb-rhea-android_ics on 1/30/2012 9PM.

    [Platform]
android_rheaberri_edn40

[BROADCOM INTERNAL NOTES]

+++ sound/caph/audio_driver/at_aud_ctl.c
@@ -238,9 +238,9 @@ int AtMaudMode(brcm_alsa_chip_t *pChip, Int32 ParamCount, Int32 *Params)
 		/* PCG and loadcal currently use Q13p2 gain format */
 	case 100:
 		{
-		int gain;
+		short gain;
 
-		gain = (int)Params[3];
+		gain = (short)Params[3];
 
 		if (Params[1] == 3) {
